[FIX] point_of_sale: preserve session closing integrity
Refunds and counts crossing zero could reverse bank balances; unused bank
methods could silently lose closing differences. Invalid monetary inputs
and concurrent cash mutations also compromised session closing.

Solution:
- Preserve account roles and POS receivables when changing payment direction.
- Post unused-method differences using the ordinary payment account defaults.
- Validate monetary inputs and serialize session cash mutations with savepoints.
- Preserve closed-session drawers and refresh live cash-control dependencies.
- Bound previous-session materialization, group closing payments once, remove
  unused helpers and deduplicate extension model loads.
- Add ledger, default-account, historical-drawer and live-transaction regressions.

Validation: 145 selected server-side tests passed with debug logging, including
settlement reconciliation; real PostgreSQL batch-lock probes and Ruff passed.
Browser and fiscal-provider flows were not exercised. Reporting assertions use
the signed-line report implementation already committed by the reporting audit.

+++ addons/pos_hr/models/pos_session.py
@@ -18,7 +18,7 @@ class PosSession(models.Model):
             data += ["hr.employee"]
         return data
 
-    def _set_opening_control_data(self, cashbox_value: int, notes: str):
+    def _set_opening_control_data(self, cashbox_value: float, notes: str):
         super()._set_opening_control_data(cashbox_value, notes)
         if author_id := self._get_message_author():
             self.message_post(
